pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "genetic-mcp"
version = "0.1.0"
description = "MCP server for genetic algorithm-based idea generation and evolution using parallel LLM workers"
authors = [{name = "andrasfe", email = ""}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "aiofiles>=23.0.0",
    "sse-starlette>=1.0.0",
    "uvicorn>=0.24.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-httpx>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
    "types-aiofiles>=23.0.0",
    "pre-commit>=3.5.0",
]

[project.scripts]
genetic-mcp = "genetic_mcp.server:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "C90", "UP", "SIM"]
ignore = ["E501", "C901"]

[tool.hatch.build.targets.wheel]
packages = ["genetic_mcp"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
exclude = [".venv/", "venv/"]

[[tool.mypy.overrides]]
module = ["mcp.*", "fastmcp.*", "sse_starlette.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = false